Homeworks/HW4/airflow-docker/requirements.txt
# Apache Airflow providers and extras
psycopg2-binary
numpy
sqlalchemy==1.4.51
pandas==1.5.0
scikit-learn
tomli
pyarrow
boto3==1.33.13
bottleneck==1.3.8
